+++ src/mixins/overflowWrap.js
@@ -1,10 +1,8 @@
-import {css} from 'styled-components';
+const overflowWrap = {
+	overflowWrap: 'break-word',
+	wordWrap: 'break-word',
 
-const overflowWrap = css`
-	overflow-wrap: break-word;
-	word-wrap: break-word;
-
-	word-break: break-word;
-`;
+	wordBreak: 'break-word',
+};
 
 export default overflowWrap;